+++ scripts/ovn_stop.sh
@@ -1,7 +1,12 @@
 #!/bin/bash
 
 source ovn_config.sh
-source minion1_args.sh
+retval=$?
+if [ $retval -ne 0 ]
+then
+  echo -e "Cannot include ovn_config.sh - maybe it is sourced from a wrong place!"
+  exit -1
+fi
 
 echo -ne "${orange}Stopping processes and deleting directories...${none}"
 sudo pkill ovs-vswitchd
@@ -9,7 +14,6 @@ sudo pkill ovsdb-server
 sudo pkill ovsdb-server
 sudo pkill ovn-northd
 sudo pkill ovn-controller
-sudo pkill ovnkube
 sudo rm -rf $OVN_PID_DIR
 sudo rm -rf $OVN_DB_FILE_DIR
 sudo rm -rf $OVN_LOG_DIR
